\Total:                   40.0 of 70.0\

+++ lex.yy.cc
@@ -1030,14 +1030,13 @@ YY_RULE_SETUP
     if(*char_buffer != '\n') {
       results.push_back(char_buffer);
       fprintf(stderr, "%s", char_buffer);
-
     }
     eof = read(pout[0], char_buffer, 1);
   }
 
   while(!results.empty()) {
     fprintf(stderr, "%s", results.back()[0]);
-    myunputc((int)results.back()[0]);
+    myunputc(results.back()[0]);
     results.pop_back();
   }
 
@@ -1077,7 +1076,7 @@ YY_RULE_SETUP
 /* Exit the shell */
 case 10:
 YY_RULE_SETUP
-#line 226 "shell.l"
+#line 225 "shell.l"
 {
   exit(0);
 }
@@ -1086,14 +1085,14 @@ YY_RULE_SETUP
 case 11:
 /* rule 11 can match eol */
 YY_RULE_SETUP
-#line 231 "shell.l"
+#line 230 "shell.l"
 {
   return NEWLINE;
 }
 	YY_BREAK
 case 12:
 YY_RULE_SETUP
-#line 235 "shell.l"
+#line 234 "shell.l"
 {
   /* Discard spaces and tabs */
 }
@@ -1101,7 +1100,7 @@ YY_RULE_SETUP
 /* Pipe */
 case 13:
 YY_RULE_SETUP
-#line 240 "shell.l"
+#line 239 "shell.l"
 {
   return PIPE;
 }
@@ -1109,7 +1108,7 @@ YY_RULE_SETUP
 /* Output redirection (stdout: 1 ) */
 case 14:
 YY_RULE_SETUP
-#line 245 "shell.l"
+#line 244 "shell.l"
 {
   return GREAT;
 }
@@ -1117,7 +1116,7 @@ YY_RULE_SETUP
 /* Input redirection */
 case 15:
 YY_RULE_SETUP
-#line 250 "shell.l"
+#line 249 "shell.l"
 {
   return LESS;
 }
@@ -1125,7 +1124,7 @@ YY_RULE_SETUP
 /* Output redirection (stderr : 2) */
 case 16:
 YY_RULE_SETUP
-#line 255 "shell.l"
+#line 254 "shell.l"
 {
   return TWOGREAT;
 }
@@ -1133,7 +1132,7 @@ YY_RULE_SETUP
 /* Output redirection (stdout and stderr : 1 and 2) */
 case 17:
 YY_RULE_SETUP
-#line 260 "shell.l"
+#line 259 "shell.l"
 {
   return GREATAMPERSAND;
 }
@@ -1141,7 +1140,7 @@ YY_RULE_SETUP
 /* Append output (stdout : 1) */
 case 18:
 YY_RULE_SETUP
-#line 265 "shell.l"
+#line 264 "shell.l"
 {
   return GREATGREAT;
 }
@@ -1149,7 +1148,7 @@ YY_RULE_SETUP
 /* Append output (stdout and stderr : 1 and 2) */
 case 19:
 YY_RULE_SETUP
-#line 270 "shell.l"
+#line 269 "shell.l"
 {
   return GREATGREATAMPERSAND;
 }
@@ -1157,7 +1156,7 @@ YY_RULE_SETUP
 /* Run process in background */
 case 20:
 YY_RULE_SETUP
-#line 275 "shell.l"
+#line 274 "shell.l"
 {
   return AMPERSAND;
 }
@@ -1168,7 +1167,7 @@ YY_RULE_SETUP
 case 21:
 /* rule 21 can match eol */
 YY_RULE_SETUP
-#line 282 "shell.l"
+#line 281 "shell.l"
 {
     /* Set up the strings for use */
     buffer.clear();
@@ -1246,7 +1245,7 @@ YY_RULE_SETUP
 /* Invalid character in input */
 case 22:
 YY_RULE_SETUP
-#line 357 "shell.l"
+#line 356 "shell.l"
 {
   /* return NOTOKEN; */
 }
@@ -1256,7 +1255,7 @@ case YY_STATE_EOF(INITIAL):
 case YY_STATE_EOF(quotes):
 case YY_STATE_EOF(manual_source):
 case YY_STATE_EOF(subshell):
-#line 362 "shell.l"
+#line 361 "shell.l"
 {
   yypop_buffer_state();
   if (!YY_CURRENT_BUFFER) {
@@ -1266,10 +1265,10 @@ case YY_STATE_EOF(subshell):
 	YY_BREAK
 case 23:
 YY_RULE_SETUP
-#line 369 "shell.l"
+#line 368 "shell.l"
 ECHO;
 	YY_BREAK
-#line 1273 "lex.yy.cc"
+#line 1272 "lex.yy.cc"
 
 	case YY_END_OF_BUFFER:
 		{
@@ -2330,4 +2329,4 @@ void yyfree (void * ptr )
 
 #define YYTABLES_NAME "yytables"
 
-#line 369 "shell.l"
+#line 368 "shell.l"
